backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.4
email-validator==2.1.2
redis==5.0.1
websockets==12.0
celery==5.3.4